style: move types in a dedicated file
